+++ base/clockwork.rb
@@ -1,0 +1,6 @@
+# Directly exit - do not raise and report a SignalException
+Signal.trap('TERM') { exit }
+
+load 'config/clockwork.rb'
+
+Clockwork.run