+++ base/clockwork.sh
@@ -4,5 +4,5 @@ cd /home/app/webapp
 
 bundle show clockwork || sleep infinity
 
-exec chpst -u app bundle exec rails runner 'load "config/clockwork.rb"; Clockwork::run' \
+exec chpst -u app bundle exec rails runner /etc/clockwork.rb \
   2>&1 |logger -t clockwork